Move `@DecoratedBy` in a dedicated package
Remove a useless field (`decorated`)
