+++ src/main/java/fr/insalyon/citi/golo/runtime/DecoratorsHelper.java
@@ -15,7 +15,7 @@
  */
 package fr.insalyon.citi.golo.runtime;
 
-import gololang.DecoratedBy;
+import gololang.annotations.DecoratedBy;
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
