+++ src/main/java/gololang/annotations/DecoratedBy.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package gololang;
+package gololang.annotations;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
